Cleaned up view of popup screens
Left aligned the back button and changed its icon color to white. Changed title on new entry popup screen. Turned off animation for TabNavigator. We can see if this helps the performance or not, but I was worried that the animation was slowing things down. I left the swipeing between screens enabled.